backend/requirements.txt
black==23.9.1
celery==5.3.4
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
mixer==7.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
